requirements.txt
Django~=3.2.12
djangorestframework~=3.14.0
django-ckeditor~=6.7.0